requirements.txt
future==0.18.3
iso8601==1.1.0
pyserial==3.5
PyYAML==6.0
